+++ packages/eslint-plugin/configs/customize.ts
@@ -13,6 +13,8 @@ type Rules = Partial<{
  * A factory function to customize the recommended config
  */
 export function customize(options: StylisticCustomizeOptions = {}): Linter.Config {
+  const defaultPluginName = plugin.meta.namespace
+
   const {
     arrowParens = false,
     blockSpacing = true,
@@ -21,27 +23,16 @@ export function customize(options: StylisticCustomizeOptions = {}): Linter.Confi
     experimental: enableExperimentalRules = false,
     indent = 2,
     jsx = true,
-    pluginName = '@stylistic',
+    pluginName = defaultPluginName,
     quoteProps = 'consistent-as-needed',
     quotes = 'single',
     semi = false,
     severity = 'error',
   } = options
 
-  let rules: Rules = {
-    '@stylistic/array-bracket-spacing': [severity, 'never'],
-    '@stylistic/arrow-parens': [severity, arrowParens ? 'always' : 'as-needed', { requireForBlockBody: true }],
-    '@stylistic/arrow-spacing': [severity, { after: true, before: true }],
-    '@stylistic/block-spacing': [severity, blockSpacing ? 'always' : 'never'],
-    '@stylistic/brace-style': [severity, braceStyle, { allowSingleLine: true }],
-    '@stylistic/comma-dangle': [severity, commaDangle],
-    '@stylistic/comma-spacing': [severity, { after: true, before: false }],
-    '@stylistic/comma-style': [severity, 'last'],
-    '@stylistic/computed-property-spacing': [severity, 'never', { enforceForClassMembers: true }],
-    '@stylistic/dot-location': [severity, 'property'],
-    '@stylistic/eol-last': severity,
-    '@stylistic/generator-star-spacing': [severity, { after: true, before: false }],
-    '@stylistic/indent': [severity, indent, {
+  const [
+    indentLevel = 2,
+    indentOptions = {
       ArrayExpression: 1,
       CallExpression: { arguments: 1 },
       flatTernaryExpressions: false,
@@ -58,10 +49,26 @@ export function customize(options: StylisticCustomizeOptions = {}): Linter.Confi
       offsetTernaryExpressions: true,
       outerIIFEBody: 1,
       SwitchCase: 1,
-      tabLength: indent === 'tab' ? 4 : indent,
+      tabLength: indentLevel === 'tab' ? 4 : indentLevel,
       VariableDeclarator: 1,
-    }],
-    '@stylistic/indent-binary-ops': [severity, indent],
+    } satisfies RuleOptions['@stylistic/indent'][1],
+  ] = Array.isArray(indent) ? indent : [indent]
+
+  let rules: Rules = {
+    '@stylistic/array-bracket-spacing': [severity, 'never'],
+    '@stylistic/arrow-parens': [severity, arrowParens ? 'always' : 'as-needed', { requireForBlockBody: true }],
+    '@stylistic/arrow-spacing': [severity, { after: true, before: true }],
+    '@stylistic/block-spacing': [severity, blockSpacing ? 'always' : 'never'],
+    '@stylistic/brace-style': [severity, braceStyle, { allowSingleLine: true }],
+    '@stylistic/comma-dangle': [severity, commaDangle],
+    '@stylistic/comma-spacing': [severity, { after: true, before: false }],
+    '@stylistic/comma-style': [severity, 'last'],
+    '@stylistic/computed-property-spacing': [severity, 'never', { enforceForClassMembers: true }],
+    '@stylistic/dot-location': [severity, 'property'],
+    '@stylistic/eol-last': severity,
+    '@stylistic/generator-star-spacing': [severity, { after: true, before: false }],
+    '@stylistic/indent': [severity, indentLevel, indentOptions],
+    '@stylistic/indent-binary-ops': [severity, indentLevel],
     '@stylistic/key-spacing': [severity, { afterColon: true, beforeColon: false }],
     '@stylistic/keyword-spacing': [severity, { after: true, before: true }],
     '@stylistic/lines-between-class-members': [severity, 'always', { exceptAfterSingleLine: true }],
@@ -99,7 +106,7 @@ export function customize(options: StylisticCustomizeOptions = {}): Linter.Confi
     '@stylistic/no-mixed-spaces-and-tabs': severity,
     '@stylistic/no-multi-spaces': severity,
     '@stylistic/no-multiple-empty-lines': [severity, { max: 1, maxBOF: 0, maxEOF: 0 }],
-    '@stylistic/no-tabs': indent === 'tab' ? 'off' : severity,
+    '@stylistic/no-tabs': indentLevel === 'tab' ? 'off' : severity,
     '@stylistic/no-trailing-spaces': severity,
     '@stylistic/no-whitespace-before-property': severity,
     '@stylistic/object-curly-spacing': [severity, 'always'],
@@ -144,7 +151,7 @@ export function customize(options: StylisticCustomizeOptions = {}): Linter.Confi
           '@stylistic/jsx-equals-spacing': severity,
           '@stylistic/jsx-first-prop-new-line': severity,
           '@stylistic/jsx-function-call-newline': [severity, 'multiline'],
-          '@stylistic/jsx-indent-props': [severity, indent],
+          '@stylistic/jsx-indent-props': [severity, indentLevel],
           '@stylistic/jsx-max-props-per-line': [severity, { maximum: 1, when: 'multiline' }],
           '@stylistic/jsx-one-expression-per-line': [severity, { allow: 'single-child' }],
           '@stylistic/jsx-quotes': severity,
@@ -190,8 +197,8 @@ export function customize(options: StylisticCustomizeOptions = {}): Linter.Confi
     }
   }
 
-  if (pluginName !== '@stylistic') {
-    const regex = /^@stylistic\//
+  if (pluginName !== defaultPluginName) {
+    const regex = new RegExp(`^${defaultPluginName}/`)
     rules = Object.fromEntries(
       Object.entries(rules!)
         .map(([ruleName, ruleConfig]) => [