+++ packages/eslint-plugin/rules/computed-property-spacing/computed-property-spacing.ts
@@ -15,13 +15,10 @@ export default createRule<RuleOptions, MessageIds>({
   name: 'computed-property-spacing',
   meta: {
     type: 'layout',
-
     docs: {
       description: 'Enforce consistent spacing inside computed property brackets',
     },
-
     fixable: 'whitespace',
-
     schema: [
       {
         type: 'string',
@@ -32,13 +29,12 @@ export default createRule<RuleOptions, MessageIds>({
         properties: {
           enforceForClassMembers: {
             type: 'boolean',
-            default: true,
           },
         },
         additionalProperties: false,
       },
     ],
-
+    defaultOptions: ['never', { enforceForClassMembers: true }],
     messages: {
       unexpectedSpaceBefore: 'There should be no space before \'{{tokenValue}}\'.',
       unexpectedSpaceAfter: 'There should be no space after \'{{tokenValue}}\'.',
@@ -47,11 +43,12 @@ export default createRule<RuleOptions, MessageIds>({
       missingSpaceAfter: 'A space is required after \'{{tokenValue}}\'.',
     },
   },
-
-  create(context) {
+  create(context, [style, options]) {
     const sourceCode = context.sourceCode
-    const propertyNameMustBeSpaced = context.options[0] === 'always' // default is "never"
-    const enforceForClassMembers = !context.options[1] || context.options[1].enforceForClassMembers
+    const propertyNameMustBeSpaced = style === 'always'
+    const {
+      enforceForClassMembers,
+    } = options!
 
     /**
      * Reports that there shouldn't be a space after the first token