+++ packages/eslint-plugin/rules/implicit-arrow-linebreak/implicit-arrow-linebreak.ts
@@ -5,35 +5,32 @@
 
 import type { Tree } from '#types'
 import type { MessageIds, RuleOptions } from './types'
-import { hasCommentsBetween, isNotOpeningParenToken, isTokenOnSameLine } from '#utils/ast'
+import { isNotOpeningParenToken, isTokenOnSameLine } from '#utils/ast'
 import { createRule } from '#utils/create-rule'
+import { safeReplaceTextBetween } from '#utils/fix'
 
 export default createRule<RuleOptions, MessageIds>({
   name: 'implicit-arrow-linebreak',
   meta: {
     type: 'layout',
-
     docs: {
       description: 'Enforce the location of arrow function bodies',
     },
-
     fixable: 'whitespace',
-
     schema: [
       {
         type: 'string',
         enum: ['beside', 'below'],
       },
     ],
+    defaultOptions: ['beside'],
     messages: {
       expected: 'Expected a linebreak before this expression.',
       unexpected: 'Expected no linebreak before this expression.',
     },
   },
-
-  create(context) {
+  create(context, [option]) {
     const sourceCode = context.sourceCode
-    const option = context.options[0] || 'beside'
 
     /**
      * Validates the location of an arrow function body
@@ -59,12 +56,7 @@ export default createRule<RuleOptions, MessageIds>({
         context.report({
           node: firstTokenOfBody,
           messageId: 'unexpected',
-          fix(fixer) {
-            if (hasCommentsBetween(sourceCode, arrowToken, firstTokenOfBody))
-              return null
-
-            return fixer.replaceTextRange([arrowToken.range[1], firstTokenOfBody.range[0]], ' ')
-          },
+          fix: safeReplaceTextBetween(sourceCode, arrowToken, firstTokenOfBody, ' '),
         })
       }
     }