+++ packages/eslint-plugin/rules/jsx-self-closing-comp/jsx-self-closing-comp.ts
@@ -1,50 +1,41 @@
-/**
- * @fileoverview Prevent extra closing tags for components without children
- * @author Yannick Croissant
- */
-
 import type { Tree } from '#types'
 import type { MessageIds, RuleOptions } from './types'
 import { isDOMComponent } from '#utils/ast/jsx'
 import { createRule } from '#utils/create-rule'
 
-const optionDefaults = { component: true, html: true }
-
-const messages = {
-  notSelfClosing: 'Empty components are self-closing',
-}
-
 export default createRule<RuleOptions, MessageIds>({
   name: 'jsx-self-closing-comp',
   meta: {
     type: 'layout',
-
     docs: {
       description: 'Disallow extra closing tags for components without children',
     },
     fixable: 'code',
-
-    messages,
-
     schema: [
       {
         type: 'object',
         properties: {
           component: {
-            default: optionDefaults.component,
             type: 'boolean',
           },
           html: {
-            default: optionDefaults.html,
             type: 'boolean',
           },
         },
         additionalProperties: false,
       },
     ],
+    defaultOptions: [{ component: true, html: true }],
+    messages: {
+      notSelfClosing: 'Empty components are self-closing',
+    },
   },
+  create(context, [configuration]) {
+    const {
+      component,
+      html,
+    } = configuration!
 
-  create(context) {
     function isComponent(node: Tree.JSXOpeningElement) {
       return (
         node.name
@@ -69,11 +60,9 @@ export default createRule<RuleOptions, MessageIds>({
     }
 
     function isShouldBeSelfClosed(node: Tree.JSXOpeningElement): node is Tree.JSXOpeningElement {
-      const configuration = Object.assign({}, optionDefaults, context.options[0])
-      return (
-        (configuration.component && isComponent(node))
-        || (configuration.html && isDOMComponent(node))
-      ) && !node.selfClosing && (childrenIsEmpty(node) || childrenIsMultilineSpaces(node))
+      return ((!!component && isComponent(node)) || (!!html && isDOMComponent(node)))
+        && !node.selfClosing
+        && (childrenIsEmpty(node) || childrenIsMultilineSpaces(node))
     }
 
     return {