+++ packages/eslint-plugin/rules/no-multi-spaces/no-multi-spaces.ts
@@ -12,13 +12,10 @@ export default createRule<RuleOptions, MessageIds>({
   name: 'no-multi-spaces',
   meta: {
     type: 'layout',
-
     docs: {
       description: 'Disallow multiple spaces',
     },
-
     fixable: 'whitespace',
-
     schema: [
       {
         type: 'object',
@@ -34,30 +31,34 @@ export default createRule<RuleOptions, MessageIds>({
           },
           ignoreEOLComments: {
             type: 'boolean',
-            default: false,
           },
           includeTabs: {
             type: 'boolean',
-            default: true,
           },
         },
         additionalProperties: false,
       },
     ],
-
+    defaultOptions: [{
+      exceptions: { Property: true, ImportAttribute: true },
+      ignoreEOLComments: false,
+      includeTabs: true,
+    }],
     messages: {
       multipleSpaces: 'Multiple spaces found before \'{{displayValue}}\'.',
     },
   },
+  create(context, [options]) {
+    const {
+      ignoreEOLComments,
+      exceptions,
+      includeTabs,
+    } = options!
 
-  create(context) {
     const sourceCode = context.sourceCode
-    const options = context.options[0] || {}
-    const ignoreEOLComments = options.ignoreEOLComments
-    const exceptions = Object.assign({ Property: true, ImportAttribute: true }, options.exceptions)
-    const hasExceptions = Object.keys(exceptions).some(key => exceptions[key])
+    const hasExceptions = Object.keys(exceptions!).some(key => exceptions![key])
 
-    const spacesRe = options.includeTabs === false ? / {2}/ : /[ \t]{2}/
+    const spacesRe = includeTabs ? /[ \t]{2}/ : / {2}/
 
     /**
      * Formats value of given comment token for error message by truncating its length.
@@ -105,7 +106,7 @@ export default createRule<RuleOptions, MessageIds>({
           if (hasExceptions) {
             const parentNode = sourceCode.getNodeByRangeIndex(rightToken.range[0] - 1)
 
-            if (parentNode && exceptions[parentNode.type])
+            if (parentNode && exceptions![parentNode.type])
               return
           }