+++ packages/eslint-plugin/rules/semi-style/semi-style.ts
@@ -76,22 +76,18 @@ export default createRule<RuleOptions, MessageIds>({
   name: 'semi-style',
   meta: {
     type: 'layout',
-
     docs: {
       description: 'Enforce location of semicolons',
     },
-
-    schema: [{ type: 'string', enum: ['last', 'first'] }],
     fixable: 'whitespace',
-
+    schema: [{ type: 'string', enum: ['last', 'first'] }],
+    defaultOptions: ['last'],
     messages: {
       expectedSemiColon: 'Expected this semicolon to be at {{pos}}.',
     },
   },
-
-  create(context) {
+  create(context, [option]) {
     const sourceCode = context.sourceCode
-    const option = context.options[0] || 'last'
 
     /**
      * Check the given semicolon token.
@@ -135,7 +131,7 @@ export default createRule<RuleOptions, MessageIds>({
         const lastToken = sourceCode.getLastToken(node)!
 
         if (isSemicolonToken(lastToken))
-          check(lastToken, option)
+          check(lastToken, option!)
       },
 
       ForStatement(node) {