+++ packages/eslint-plugin/rules/space-unary-ops/space-unary-ops.ts
@@ -12,24 +12,19 @@ export default createRule<RuleOptions, MessageIds>({
   name: 'space-unary-ops',
   meta: {
     type: 'layout',
-
     docs: {
       description: 'Enforce consistent spacing before or after unary operators',
     },
-
     fixable: 'whitespace',
-
     schema: [
       {
         type: 'object',
         properties: {
           words: {
             type: 'boolean',
-            default: true,
           },
           nonwords: {
             type: 'boolean',
-            default: false,
           },
           overrides: {
             type: 'object',
@@ -41,6 +36,7 @@ export default createRule<RuleOptions, MessageIds>({
         additionalProperties: false,
       },
     ],
+    defaultOptions: [{ words: true, nonwords: false }],
     messages: {
       unexpectedBefore: 'Unexpected space before unary operator \'{{operator}}\'.',
       unexpectedAfter: 'Unexpected space after unary operator \'{{operator}}\'.',
@@ -50,9 +46,12 @@ export default createRule<RuleOptions, MessageIds>({
       requireBefore: 'Space is required before unary operator \'{{operator}}\'.',
     },
   },
-
-  create(context) {
-    const options = context.options[0] || { words: true, nonwords: false }
+  create(context, [options]) {
+    const {
+      words,
+      nonwords,
+      overrides = {},
+    } = options!
 
     const sourceCode = context.sourceCode
 
@@ -63,24 +62,6 @@ export default createRule<RuleOptions, MessageIds>({
      */
     function isFirstBangInBangBangExpression(node: ASTNode) {
       return node && node.type === 'UnaryExpression' && node.argument && node.argument.type === 'UnaryExpression' && node.argument.operator === '!'
-    }
-
-    /**
-     * Checks if an override exists for a given operator.
-     * @param operator Operator
-     * @returns Whether or not an override has been provided for the operator
-     */
-    function overrideExistsForOperator(operator: string) {
-      return options.overrides && Object.hasOwn(options.overrides, operator)
-    }
-
-    /**
-     * Gets the value that the override was set to for this operator
-     * @param operator Operator
-     * @returns Whether or not an override enforces a space with this operator
-     */
-    function overrideEnforcesSpaces(operator: string) {
-      return options.overrides?.[operator]
     }
 
     /**
@@ -137,7 +118,7 @@ export default createRule<RuleOptions, MessageIds>({
      * @param word The word to be used for reporting
      */
     function checkUnaryWordOperatorForSpaces(node: ASTNode, firstToken: Token, secondToken: Token, word: string) {
-      const shouldHaveSpace = overrideExistsForOperator(word) ? overrideEnforcesSpaces(word) : options.words
+      const shouldHaveSpace = overrides[word] ?? words
 
       if (shouldHaveSpace) {
         verifyWordHasSpaces(node, firstToken, secondToken, word)
@@ -156,7 +137,7 @@ export default createRule<RuleOptions, MessageIds>({
       const operatorToken = sourceCode.getLastToken(node, token => token.value === operator)!
       const prefixToken = sourceCode.getTokenBefore(operatorToken)!
 
-      const shouldHaveSpace = overrideExistsForOperator(operator) ? overrideEnforcesSpaces(operator) : options.nonwords
+      const shouldHaveSpace = overrides[operator] ?? nonwords
 
       if (shouldHaveSpace) {
         verifyNonWordsHaveSpaces(node, prefixToken, operatorToken)
@@ -309,7 +290,7 @@ export default createRule<RuleOptions, MessageIds>({
 
       const operator = ('prefix' in node && node.prefix) ? tokens[0].value : tokens[1].value
 
-      const shouldHaveSpace = overrideExistsForOperator(operator) ? overrideEnforcesSpaces(operator) : options.nonwords
+      const shouldHaveSpace = overrides[operator] ?? nonwords
 
       if (shouldHaveSpace) {
         verifyNonWordsHaveSpaces(node, firstToken, secondToken)